fixup

+++ tests/perft-final.cpp
@@ -6,7 +6,7 @@
 #include "../src/utils/helpers.h"
 
 template<Color Us>
-unsigned long long perft_node_count(Position& p, unsigned int depth) {
+u64 perft_node_count(Position& p, unsigned int depth) {
 	unsigned long long nodes = 0;
 
 	MoveList<Us, ALL> list(p);
@@ -19,7 +19,7 @@ unsigned long long perft_node_count(Position& p, unsigned int depth) {
 	return nodes;
 }
 
-unsigned long long test_perft_node_count(const std::string& fen, int depth) {
+u64 test_perft_node_count(const std::string& fen, int depth) {
 	Position p(fen);
 	if (p.side == WHITE) {
 		return perft_node_count<WHITE>(p, depth);
@@ -31,39 +31,39 @@ unsigned long long test_perft_node_count(const std::string& fen, int depth) {
 TEST_SUITE_BEGIN("perft-final");
 
 TEST_CASE("PerftDepthSixDefaultFen") {
-			CHECK_EQ(test_perft_node_count(START_FEN, 6), 119060324);
+	CHECK_EQ(test_perft_node_count(START_FEN, 6), 119060324);
 }
 
 TEST_CASE("PerftDepthFiveKiwipete") {
-			CHECK_EQ(test_perft_node_count(KIWIPETE_FEN, 5), 193690690);
+	CHECK_EQ(test_perft_node_count(KIWIPETE_FEN, 5), 193690690);
 }
 
 TEST_CASE("PerftTestPosition") {
-			CHECK_EQ(test_perft_node_count("rn2kbnr/pp2pppp/2p5/8/5B2/N2P1B1P/PP3PP1/R4RK1 w kq - 0 13", 5), 18292958);
+	CHECK_EQ(test_perft_node_count("rn2kbnr/pp2pppp/2p5/8/5B2/N2P1B1P/PP3PP1/R4RK1 w kq - 0 13", 5), 18292958);
 }
 
 TEST_CASE("PerftTestPosition2") {
-			CHECK_EQ(test_perft_node_count("8/8/8/3p4/4pn1N/6p1/8/5K1k w - - 10 73", 8), 24490963);
+	CHECK_EQ(test_perft_node_count("8/8/8/3p4/4pn1N/6p1/8/5K1k w - - 10 73", 8), 24490963);
 }
 
 TEST_CASE("PerftTestPosition5") {
-			CHECK_EQ(test_perft_node_count("rnb1kbnr/ppp1pppp/4q3/3pP3/8/4K3/PPPP1PPP/RNBQ1BNR w - d6 0 1", 1), 34);
+	CHECK_EQ(test_perft_node_count("rnb1kbnr/ppp1pppp/4q3/3pP3/8/4K3/PPPP1PPP/RNBQ1BNR w - d6 0 1", 1), 34);
 }
 
 TEST_CASE("PerftTestPosition6") {
-			CHECK_EQ(test_perft_node_count("rnb1kbnr/ppp1pppp/4q3/3pP3/8/5K2/PPPP1PPP/RNBQ1BNR w - d6 0 1", 1), 28);
+	CHECK_EQ(test_perft_node_count("rnb1kbnr/ppp1pppp/4q3/3pP3/8/5K2/PPPP1PPP/RNBQ1BNR w - d6 0 1", 1), 28);
 }
 
 TEST_CASE("PerftTestPosition7") {
-			CHECK_EQ(test_perft_node_count("rnb1kbnr/ppq1pppp/p7/3pP3/8/6K1/PPPP1PPP/RNBQ1BNR w - d6 0 1", 1), 31);
+	CHECK_EQ(test_perft_node_count("rnb1kbnr/ppq1pppp/p7/3pP3/8/6K1/PPPP1PPP/RNBQ1BNR w - d6 0 1", 1), 31);
 }
 
 TEST_CASE("PerftTestPosition3") {
-			CHECK_EQ(test_perft_node_count("8/8/8/3p4/5n2/4p3/6pk/4KN2 b - - 3 76", 1), 7);
+	CHECK_EQ(test_perft_node_count("8/8/8/3p4/5n2/4p3/6pk/4KN2 b - - 3 76", 1), 7);
 }
 
 TEST_CASE("PerftTestPositionKnightCheck"){
-			CHECK_EQ(test_perft_node_count("5n2/3KPk2/8/3p1N2/3P4/8/8/8 w - - 17 102", 8), 80433958);
+	CHECK_EQ(test_perft_node_count("5n2/3KPk2/8/3p1N2/3P4/8/8/8 w - - 17 102", 8), 80433958);
 }
 
 TEST_CASE("perft-all") {
@@ -81,8 +81,7 @@ TEST_CASE("perft-all") {
 			std::string expected = split_perft[i];
 			int depth = static_cast<i32>(i);
 			int expected_node_count = std::stoi(split(expected, " ")[2]);
-			//std::cout << "Running fen: " << fen << " depth " << depth << " expected node count " << expected_node_count << std::endl;
-					CHECK_EQ(test_perft_node_count(fen, depth), expected_node_count);
+			CHECK_EQ(test_perft_node_count(fen, depth), expected_node_count);
 			total_nodes += expected_node_count;
 		}
 	}
